Changed Models and Storage Functions
