+++ backend/models/auth.ts
@@ -1,0 +1,4 @@
+export interface JWTBlacklistItem {
+    id: string;
+    jwt: string;
+}
